Move class Filesystem into Foundation\Kernel namespace
