+++ packages/framework/tests/Feature/FilesystemFacadeTest.php
@@ -13,7 +13,7 @@ use Illuminate\Support\LazyCollection;
 
 /**
  * @covers \Hyde\Facades\Filesystem
- * @covers \Hyde\Foundation\Filesystem
+ * @covers \Hyde\Foundation\Kernel\Filesystem
  * @covers \Hyde\Framework\Concerns\Internal\ForwardsIlluminateFilesystem
  */
 class FilesystemFacadeTest extends TestCase
